Provide endpoint for 'versions.php', which returns a json object.
This is an attempt to improve backwards compatibility for old versions of MIPSym.
When old versions of MIPSym try to connect to MIPSym.com, they will try to access
`MIPSym.com/versions.php`, and they will expect to receive a json object that
includes the version number of the most recent version of MIPSym. This file
provides that functionality.

`versions.php` was changed to a static `versions.json` file when the website moved
to GitHub, because you can't serve php on GitHub Pages. The new `versions.php`
file will not always be up to date, but it should convince anyone with any
version of MIPSym older than 2.05.1519 that there's a new version available.
When they update to a new version, their copy of MIPSym should start checking
`versions.json` instead of `versions.php`.

It is possible that `versions.php` won't help anyone at all. The Windows version
of MIPSym lacks the proper dlls for dealing with `https` websites, and
will fail to connect with the `MIPSym.com/versions.php` endpoint under any
circumstance. I am uncertain what the Windows `maintenancetool.exe` will do.
Regardless, `versions.php` needs to be here, even if it isn't a real php file.